+++ app/frontend/package.json
@@ -17,7 +17,6 @@
 	"devDependencies": {
 		"@eslint/js": "^9.34.0",
 		"@internationalized/date": "^3.9.0",
-		"@lucide/svelte": "^0.515.0",
 		"@sveltejs/adapter-node": "^5.2.13",
 		"@sveltejs/kit": "^2.22.0",
 		"@sveltejs/vite-plugin-svelte": "^6.0.0",
@@ -49,10 +48,10 @@
 		"tailwindcss": "^4.0.0",
 		"tw-animate-css": "^1.3.8",
 		"typescript": "^5.0.0",
-		"vite": "^7.0.4",
-		"vite-plugin-lucide-preprocess": "^1.4.3"
+		"vite": "^7.0.4"
 	},
 	"dependencies": {
+		"@lucide/svelte": "^0.544.0",
 		"@types/d3-array": "^3.2.1",
 		"@types/d3-scale": "^4.0.9",
 		"@types/d3-shape": "^3.1.7",
