+++ app/frontend/src/app.d.ts
@@ -1,3 +1,4 @@
+import 'unplugin-icons/types/';
 // See https://svelte.dev/docs/kit/types#app.d.ts
 // for information about these interfaces
 declare global {
