+++ app/frontend/src/lib/components/features/fuel/FuelLogTab.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import TabContainer from '$appui/TabContainer.svelte';
-	import FuelLogList from '$lib/components/features/dashboard/fuel/FuelLogList.svelte';
+	import FuelLogList from './FuelLogList.svelte';
 
 	let { vehicleId } = $props();
 </script>
