+++ app/frontend/src/lib/components/features/fuel/FuelLogTable.svelte
@@ -2,7 +2,6 @@
 	import { getCoreRowModel, type ColumnDef } from '@tanstack/table-core';
 	import { createSvelteTable, FlexRender } from '$lib/components/ui/data-table/index.js';
 	import * as Table from '$lib/components/ui/table/index.js';
-	import { Calendar1 } from '@lucide/svelte';
 
 	type DataTableProps<TData, TValue> = {
 		columns: ColumnDef<TData, TValue>[];
