+++ app/frontend/src/lib/components/features/insurance/InsuranceTab.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import TabContainer from '$appui/TabContainer.svelte';
-	import InsuranceDetailsList from '$lib/components/features/dashboard/insurance/InsuranceList.svelte';
+	import InsuranceDetailsList from './InsuranceList.svelte';
 
 	let { vehicleId } = $props();
 </script>
