+++ app/frontend/src/lib/components/features/maintenance/MaintenenceLogTab.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import TabContainer from '$appui/TabContainer.svelte';
-	import MaintenanceLogList from '$lib/components/features/dashboard/maintenance/MaintenanceLogList.svelte';
+	import MaintenanceLogList from './MaintenanceLogList.svelte';
 
 	let { vehicleId } = $props();
 </script>
