+++ app/frontend/src/lib/components/features/overview/ChartCard.svelte
