+++ app/frontend/src/lib/components/features/pullution/PollutionCertificateList.svelte
@@ -2,12 +2,12 @@
 	import { onMount } from 'svelte';
 	import { browser } from '$app/environment';
 	import { env } from '$env/dynamic/public';
-	import { FileText, Calendar, MapPin, Trash2, BadgeCheck } from '@lucide/svelte';
-	import { formatDate } from '$utils/formatting';
+	import { FileText, Calendar, MapPin, Trash2, BadgeCheck } from '@lucide/svelte/icons';
+	import { formatDate } from '$helper/formatting';
 	import { Jumper } from 'svelte-loading-spinners';
 	import IconButton from '$appui/IconButton.svelte';
 	import DeleteConfirmation from '$appui/DeleteConfirmation.svelte';
-	import { getApiUrl } from '$utils/api';
+	import { getApiUrl } from '$helper/api';
 
 	let { vehicleId } = $props();
