+++ app/frontend/src/lib/components/features/vehicle/VehicleCard.svelte
@@ -1,4 +1,6 @@
 <script lang="ts">
+	import LicensePlate from './LicensePlate.svelte';
+
 	import {
 		Gauge,
 		Pencil,
@@ -10,9 +12,10 @@
 		CircleDotDashed,
 		ShieldCheck,
 		BadgeAlert,
-		ShieldAlert
-	} from '@lucide/svelte';
-	import { formatDistance } from '$utils/formatting';
+		ShieldAlert,
+		IdCard
+	} from '@lucide/svelte/icons';
+	import { formatDistance } from '$helper/formatting';
 	import { vehicleModelStore, vehiclesStore } from '$stores/vehicle';
 	import { maintenanceModelStore } from '$stores/maintenance';
 	import { fuelLogModelStore } from '$stores/fuel-log';
@@ -25,11 +28,12 @@
 	import * as Card from '$lib/components/ui/card';
 	import Badge from '$lib/components/ui/badge/badge.svelte';
 	import IconWithTooltip from '$appui/IconWithTooltip.svelte';
+	import LabelWithIcon from '$lib/components/ui/app/LabelWithIcon.svelte';
 
-	const { vehicle, updateCallback, isSelected = false } = $props();
+	const { vehicle, updateCallback, onclick, onkeydown, isSelected = false } = $props();
 	let deleteDialog = $state(false);
 
-	async function deleteVehicle(vehicleId: string) {
+	const deleteVehicle = async (vehicleId: string) => {
 		try {
 			const response = await fetch(`${env.PUBLIC_API_BASE_URL || ''}/api/vehicles/${vehicleId}`, {
 				method: 'DELETE',
@@ -49,7 +53,7 @@
 			console.log(e);
 			alert('Failed to connect to the server.');
 		}
-	}
+	};
 
 	const fetchVehicles = () => {
 		if (browser) {
@@ -63,126 +67,124 @@
 		vehicle.imageUrl || 'https://www.v3cars.com/media/model-imgs/91-92-062100-daytona-grey.webp';
 </script>
 
-<Card.Root
-	class={`hover:border-primary h-full w-xs cursor-pointer gap-2 rounded-2xl border-2 p-0 pb-4 transition-all duration-300 ease-in-out ${isSelected ? 'border-primary' : 'border-transparent'}`}
->
-	<Card.Header class="relative h-42 overflow-hidden p-0 ">
-		<div class="w-full">
-			<img src={imageUrl} alt="car" class="rounded-t-xl object-cover opacity-30 dark:opacity-30" />
-		</div>
-		<div class="absolute inset-0 flex flex-col justify-between p-4">
-			<div class="flex flex-col">
-				<div class="flex items-center justify-between">
-					<div class="flex items-center gap-2">
-						<span class="text-2xl font-bold">{vehicle.make} {vehicle.model}</span>
-					</div>
-					<p class="flex items-center gap-2">
-						{#if vehicle.color}
-							<Badge
-								variant="outline"
-								class="m-1 h-5 w-8"
-								style={`background-color: ${vehicle.color}`}
-							></Badge>
-						{:else}
-							<span>-</span>
-						{/if}
-					</p>
-				</div>
-				<div
-					class="mono text-accent-foreground/50 mt-2 flex flex-row justify-between text-sm font-medium"
-				>
-					{vehicle.vin ? vehicle.vin : '-'}
-					<p class="mono text-foreground flex items-center gap-2">
-						<Gauge class="h-5 w-5" />
-						{vehicle.odometer ? formatDistance(vehicle.odometer) : '-'}
-					</p>
-				</div>
+<div tabindex="0" role="button" {onclick} {onkeydown}>
+	<Card.Root
+		class={`hover:border-primary h-full w-xs cursor-pointer gap-2 rounded-2xl border-2 p-0 pb-4 transition-all duration-300 ease-in-out ${isSelected ? 'border-primary' : 'border-transparent'}`}
+	>
+		<Card.Header class="relative h-38 overflow-hidden p-0 ">
+			<div class="w-full">
+				<img
+					src={imageUrl}
+					alt="car"
+					class="rounded-t-xl object-cover opacity-30 dark:opacity-30"
+				/>
 			</div>
+			<div class="absolute inset-0 flex flex-col justify-between p-4">
+				<div class="flex flex-col">
+					<div class="flex items-center justify-between">
+						<div class="flex items-center gap-2">
+							<span class="text-2xl font-bold">{vehicle.make} {vehicle.model}</span>
+						</div>
+						<p class="flex items-center gap-2">
+							{#if vehicle.color}
+								<Badge class="m-1 h-5 w-8" style={`background-color: ${vehicle.color}`} />
+							{/if}
+						</p>
+					</div>
+					<div class="mt-2 flex flex-row justify-between text-sm font-medium">
+						<LabelWithIcon
+							icon={IdCard}
+							iconClass="h-5 w-5"
+							style="mono text-zinc-600 flex items-center gap-2"
+							label={vehicle.vin ? vehicle.vin : '-'}
+						/>
+						<LabelWithIcon
+							icon={Gauge}
+							iconClass="h-5 w-5"
+							style="mono text-foreground flex items-center gap-2"
+							label={vehicle.odometer ? formatDistance(vehicle.odometer) : '-'}
+						/>
+					</div>
+				</div>
 
-			<div class="flex flex-row justify-between">
-				<IconWithTooltip
-					icon={vehicle.insuranceStatus === 'Active' ? ShieldCheck : ShieldAlert}
-					className={vehicle.insuranceStatus === 'Active'
-						? 'dark:text-green-700 text-green-500'
-						: 'dark:text-rose-600 text-rose-500'}
-					tooltip={`Insurance is ${vehicle.insuranceStatus}`}
-					side="right"
-				/>
-				<IconWithTooltip
-					icon={vehicle.puccStatus === 'Active' ? BadgeCheck : BadgeAlert}
-					className={vehicle.puccStatus === 'Active'
-						? 'dark:text-green-700 text-green-500'
-						: 'dark:text-rose-600 text-rose-500'}
-					tooltip={`Pollution Status is ${vehicle.puccStatus}`}
-					side="left"
-				/>
-			</div>
-		</div>
-	</Card.Header>
-	<Card.Content class="px-4">
-		<div class="flex items-center justify-between">
-			<div class="border-secondary inline-block rounded-md border-4 text-xl font-bold">
-				<div class="flex flex-row items-center gap-4">
-					<div class="bg-secondary flex h-8 w-8 items-center justify-center">
-						<CircleDotDashed class="h-5 w-5" />
-					</div>
-					<p class="mono pe-4 font-bold">{vehicle.licensePlate}</p>
+				<div class="flex flex-row justify-between">
+					<IconWithTooltip
+						icon={vehicle.insuranceStatus === 'Active' ? ShieldCheck : ShieldAlert}
+						className={vehicle.insuranceStatus === 'Active'
+							? 'dark:text-green-600 text-green-800'
+							: 'dark:text-rose-600 text-rose-800'}
+						tooltip={`Insurance is ${vehicle.insuranceStatus}`}
+						side="right"
+					/>
+					<IconWithTooltip
+						icon={vehicle.puccStatus === 'Active' ? BadgeCheck : BadgeAlert}
+						className={vehicle.puccStatus === 'Active'
+							? 'dark:text-green-600 text-green-800'
+							: 'dark:text-rose-600 text-rose-800'}
+						tooltip={`Pollution Status is ${vehicle.puccStatus}`}
+						side="left"
+					/>
 				</div>
 			</div>
-			<Badge class="text-zinc-100">{vehicle.year}</Badge>
-		</div>
-	</Card.Content>
-	<Card.Footer class="px-3 pt-4">
-		<div class="flex w-full justify-between">
-			<div class="flex justify-start">
-				<IconButton
-					buttonStyles="hover:bg-green-100 dark:hover:bg-green-700"
-					iconStyles="text-green-500 hover:text-green-600 dark:text-green-400 dark:hover:text-green-200"
-					icon={Fuel}
-					onclick={() => fuelLogModelStore.show(vehicle.id, null, false, updateCallback)}
-					ariaLabel="Log fuel refill"
-				/>
-				<IconButton
-					buttonStyles="hover:bg-amber-100 dark:hover:bg-amber-700"
-					iconStyles="text-amber-500 hover:text-amber-600 dark:text-amber-400 dark:hover:text-amber-200"
-					icon={Wrench}
-					onclick={() => maintenanceModelStore.show(vehicle.id, null, false, updateCallback)}
-					ariaLabel="Maintenence"
-				/>
-				<IconButton
-					buttonStyles="hover:bg-sky-100 dark:hover:bg-sky-700"
-					iconStyles="text-sky-500 hover:text-sky-600 dark:text-sky-400 dark:hover:text-sky-200"
-					icon={Shield}
-					onclick={() => insuranceModelStore.show(vehicle.id, null, false, updateCallback)}
-					ariaLabel="Insurance"
-				/>
-				<IconButton
-					buttonStyles="hover:bg-fuchsia-100 dark:hover:bg-fuchsia-700"
-					iconStyles="text-fuchsia-500 hover:text-fuchsia-600 dark:text-fuchsia-400 dark:hover:text-fuchsia-200"
-					icon={BadgeCheck}
-					onclick={() => puccModelStore.show(vehicle.id, null, false, updateCallback)}
-					ariaLabel="Pollution Certificate"
-				/>
+		</Card.Header>
+		<Card.Content class="px-4">
+			<div class="flex items-center justify-between">
+				<LicensePlate registrationNumber={vehicle.licensePlate} />
+				<Badge variant="outline" class="text-sm">{vehicle.year}</Badge>
 			</div>
-			<div class="flex justify-end gap-2">
-				<IconButton
-					buttonStyles="hover:bg-gray-200 dark:hover:bg-gray-700"
-					iconStyles="text-gray-600 dark:text-gray-100 hover:text-sky-500"
-					icon={Pencil}
-					onclick={() => {
-						vehicleModelStore.show(vehicle, true);
-					}}
-					ariaLabel="Edit"
-				/>
-				<IconButton
-					buttonStyles="hover:bg-gray-200 dark:hover:bg-gray-700"
-					iconStyles="text-gray-600 dark:text-gray-100 hover:text-red-500"
-					icon={Trash2}
-					onclick={() => (deleteDialog = true)}
-					ariaLabel="Delete"
-				/>
+		</Card.Content>
+		<Card.Footer class="px-3">
+			<div class="flex w-full justify-between">
+				<div class="flex justify-start">
+					<IconButton
+						buttonStyles="hover:bg-green-100 dark:hover:bg-green-700"
+						iconStyles="text-green-500 hover:text-green-600 dark:text-green-400 dark:hover:text-green-200"
+						icon={Fuel}
+						onclick={() => fuelLogModelStore.show(vehicle.id, null, false, updateCallback)}
+						ariaLabel="Log fuel refill"
+					/>
+					<IconButton
+						buttonStyles="hover:bg-amber-100 dark:hover:bg-amber-700"
+						iconStyles="text-amber-500 hover:text-amber-600 dark:text-amber-400 dark:hover:text-amber-200"
+						icon={Wrench}
+						onclick={() => maintenanceModelStore.show(vehicle.id, null, false, updateCallback)}
+						ariaLabel="Maintenence"
+					/>
+					<IconButton
+						buttonStyles="hover:bg-sky-100 dark:hover:bg-sky-700"
+						iconStyles="text-sky-500 hover:text-sky-600 dark:text-sky-400 dark:hover:text-sky-200"
+						icon={Shield}
+						onclick={() => insuranceModelStore.show(vehicle.id, null, false, updateCallback)}
+						ariaLabel="Insurance"
+					/>
+					<IconButton
+						buttonStyles="hover:bg-fuchsia-100 dark:hover:bg-fuchsia-700"
+						iconStyles="text-fuchsia-500 hover:text-fuchsia-600 dark:text-fuchsia-400 dark:hover:text-fuchsia-200"
+						icon={BadgeCheck}
+						onclick={() => puccModelStore.show(vehicle.id, null, false, updateCallback)}
+						ariaLabel="Pollution Certificate"
+					/>
+				</div>
+				<div class="flex justify-end gap-2">
+					<IconButton
+						buttonStyles="hover:bg-gray-200 dark:hover:bg-gray-700"
+						iconStyles="text-gray-600 dark:text-gray-100 hover:text-sky-500"
+						icon={Pencil}
+						onclick={() => {
+							vehicleModelStore.show(vehicle, true);
+						}}
+						ariaLabel="Edit"
+					/>
+					<IconButton
+						buttonStyles="hover:bg-gray-200 dark:hover:bg-gray-700"
+						iconStyles="text-gray-600 dark:text-gray-100 hover:text-red-500"
+						icon={Trash2}
+						onclick={() => (deleteDialog = true)}
+						ariaLabel="Delete"
+					/>
+				</div>
 			</div>
-		</div>
-	</Card.Footer>
-</Card.Root>
+		</Card.Footer>
+	</Card.Root>
+</div>
 <DeleteConfirmation onConfirm={() => deleteVehicle(vehicle.id)} bind:open={deleteDialog} />
