+++ app/frontend/src/lib/components/features/vehicle/VehicleList.svelte
@@ -1,38 +1,68 @@
 <script lang="ts">
 	import { vehiclesStore } from '$stores/vehicle';
 	import VehicleCard from './VehicleCard.svelte';
-	import ScrollArea from '$lib/components/ui/scroll-area/scroll-area.svelte';
+	import { ScrollArea } from '$lib/components/ui/scroll-area';
+	import LabelWithIcon from '$lib/components/ui/app/LabelWithIcon.svelte';
+	import { CircleAlert, CircleSlash2 } from '@lucide/svelte';
+	import type { Vehicle } from '$lib/types/vehicle';
+	import { Jumper } from 'svelte-loading-spinners';
 
-	let { vehicles, selectedVehicleId, updateCallback } = $props();
+	let { selectedVehicleId = $bindable(), updateCallback } = $props();
+	let vehicles = $state<Vehicle[]>([]);
+	let loading = $state(true);
+	let error = $state('');
 
-	function selectVehicle(vehicleId: string) {
+	vehiclesStore.subscribe((data) => {
+		vehicles = data.vehicles;
+		loading = data.loading;
+		error = data.error;
+		selectedVehicleId = data.selectedVehicleId;
+		if (vehicles.length > 0) {
+			selectedVehicleId = selectedVehicleId || vehicles[0].id;
+		} else {
+			selectedVehicleId = undefined;
+		}
+	});
+
+	const selectVehicle = (vehicleId: string) => {
 		vehiclesStore.selectVehicle(vehicleId);
-	}
+	};
 </script>
+
+{#if loading}
+	<div class="flex items-center justify-center">
+		<Jumper size="40" duration="2s" />
+	</div>
+{:else if error}
+	<LabelWithIcon
+		icon={CircleAlert}
+		iconClass="h-6 w-6"
+		style="flex min-h-98 items-center justify-center gap-4 text-xl"
+		label={error}
+	/>
+{/if}
 
 {#if vehicles.length > 0}
 	<ScrollArea class="w-full whitespace-nowrap" orientation="horizontal">
 		<div class="m-4 flex gap-4">
 			{#each vehicles as vehicle (vehicle.id)}
-				<div
-					tabindex="0"
-					role="button"
+				<VehicleCard
+					{vehicle}
+					{updateCallback}
+					isSelected={selectedVehicleId === vehicle.id}
 					onclick={() => selectVehicle(vehicle.id)}
-					onkeydown={(e) => {
+					onkeydown={(e: { key: string }) => {
 						if (e.key === 'Enter' || e.key === ' ') selectVehicle(vehicle.id);
 					}}
-				>
-					<VehicleCard {vehicle} {updateCallback} isSelected={selectedVehicleId === vehicle.id} />
-				</div>
+				/>
 			{/each}
 		</div>
 	</ScrollArea>
 {:else}
-	<!-- <div class="flex items-center justify-center"> -->
-	<div class="flex min-h-48 items-center justify-center gap-10 bg-gray-100 dark:bg-gray-900">
-		<p class="text-lg text-gray-600 dark:text-gray-100">
-			No vehicles found. Add your first vehicle to begin.
-		</p>
-	</div>
-	<!-- </div> -->
+	<LabelWithIcon
+		icon={CircleSlash2}
+		iconClass="h-6 w-6"
+		style="flex min-h-98 items-center justify-center gap-4 text-xl"
+		label="It's empty here. Please add your first vehicle to begin."
+	/>
 {/if}
