+++ app/frontend/src/lib/components/layout/DashboardTabs.svelte
@@ -1,0 +1,64 @@
+<script lang="ts">
+	import * as Tabs from '$lib/components/ui/tabs';
+	import OverviewTab from '../features/overview/OverviewTab.svelte';
+	import FuelLogTab from '../features/fuel/FuelLogTab.svelte';
+	import MaintenenceLogTab from '../features/maintenance/MaintenenceLogTab.svelte';
+	import InsuranceTab from '../features/insurance/InsuranceTab.svelte';
+	import PollutionTab from '../features/pullution/PollutionTab.svelte';
+	import { BadgeInfo, Fuel, Shield, SquareKanban, Wrench } from '@lucide/svelte';
+	import LabelWithIcon from '../ui/app/LabelWithIcon.svelte';
+
+	let { vehicleId = $bindable() } = $props();
+	const tabs: {
+		name: string;
+		id: string;
+		Component: any;
+		Icon?: any;
+	}[] = [
+		{
+			name: 'Overview',
+			id: 'overview',
+			Component: OverviewTab,
+			Icon: SquareKanban
+		},
+		{
+			name: 'Fuel Logs',
+			id: 'fuel',
+			Component: FuelLogTab,
+			Icon: Fuel
+		},
+		{
+			name: 'Maintenance',
+			id: 'maintenance',
+			Component: MaintenenceLogTab,
+			Icon: Wrench
+		},
+		{
+			name: 'Insurance',
+			id: 'insurance',
+			Component: InsuranceTab,
+			Icon: Shield
+		},
+		{
+			name: 'Pollution Certificate',
+			id: 'pollution',
+			Component: PollutionTab,
+			Icon: BadgeInfo
+		}
+	];
+</script>
+
+<Tabs.Root value={tabs[0].id} class="w-full">
+	<Tabs.List class="w-full">
+		{#each tabs as tab}
+			<Tabs.Trigger value={tab.id}>
+				<LabelWithIcon icon={tab.Icon} label={tab.name} />
+			</Tabs.Trigger>
+		{/each}
+	</Tabs.List>
+	{#each tabs as { id, Component } (id)}
+		<Tabs.Content class=" w-full" value={id}>
+			<Component {vehicleId} />
+		</Tabs.Content>
+	{/each}
+</Tabs.Root>
