+++ app/frontend/src/lib/components/ui/app/LabelWithIcon.svelte
@@ -1,10 +1,13 @@
 <script lang="ts">
 	import { cn } from '$lib/utils';
 
-	let { icon: Icon, label, iconClass = '', labelClass = '', style = '' } = $props();
+	let { icon: Icon, label = '', children = null, iconClass = '', style = '' } = $props();
 </script>
 
 <div class={cn('flex flex-row items-center gap-2', style)}>
 	<Icon class={iconClass} />
-	<span class={labelClass}>{label}</span>
+	<span>{label}</span>
+	{#if children}
+		{@render children()}
+	{/if}
 </div>
