+++ app/frontend/src/lib/components/ui/app/TabContainer.svelte
@@ -2,8 +2,8 @@
 	const { title, children } = $props();
 </script>
 
-<div class="bg-secondary rounded-lg p-4" role="tabpanel" aria-labelledby="dashboard-tab">
-	<h2 class="mb-6 text-2xl font-bold">
+<div class="bg-secondary rounded-md p-8" role="tabpanel" aria-labelledby="dashboard-tab">
+	<h2 class="mb-6 text-2xl font-bold underline underline-offset-8">
 		{title}
 	</h2>
 	{@render children()}
