+++ app/frontend/src/lib/components/ui/app/ThemeToggle.svelte
@@ -3,10 +3,10 @@
 	import MoonIcon from '@lucide/svelte/icons/moon';
 
 	import { toggleMode } from 'mode-watcher';
-	import { Button } from '$lib/components/ui/button/index.js';
+	import { Button } from '$lib/components/ui/button';
 </script>
 
-<Button onclick={toggleMode} variant="outline" size="icon">
+<Button onclick={toggleMode} variant="ghost" size="icon">
 	<SunIcon
 		class="h-[1.2rem] w-[1.2rem] scale-100 rotate-0 !transition-all dark:scale-0 dark:-rotate-90"
 	/>
