+++ app/frontend/src/lib/components/ui/input-otp/input-otp-separator.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import type { HTMLAttributes } from 'svelte/elements';
 	import type { WithElementRef } from '$lib/utils';
-	import { DotIcon } from '@lucide/svelte';
+	import { DotIcon } from '@lucide/svelte/icons';
 
 	let {
 		ref = $bindable(null),
