+++ app/frontend/src/lib/helper/formatting.ts
