+++ app/frontend/src/lib/stores/config.ts
@@ -1,6 +1,6 @@
 import { writable } from 'svelte/store';
 import { browser } from '$app/environment';
-import { getApiUrl } from '$utils/api';
+import { getApiUrl } from '$helper/api';
 
 export interface Config {
 	key: string;
